Private feature basic password protection implemented. Material dialog third party library migrated to newest version. Support design library and appcompat updated to latest version and black list items issue fixed by creation custom choice indicator.

+++ MyApplication/app/src/main/java/com/gcode/notes/ui/callbacks/bin/DeleteNotePermanentlyCallback.java
@@ -1,12 +1,15 @@
 package com.gcode.notes.ui.callbacks.bin;
 
+import android.support.annotation.NonNull;
+
+import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.gcode.notes.adapters.MainAdapter;
 import com.gcode.notes.data.base.ContentBase;
 import com.gcode.notes.extras.MyDebugger;
 import com.gcode.notes.notes.MyApplication;
 
-public class DeleteNotePermanentlyCallback extends MaterialDialog.ButtonCallback {
+public class DeleteNotePermanentlyCallback implements MaterialDialog.SingleButtonCallback {
     MainAdapter mAdapter;
     int mPosition;
     ContentBase mNote;
@@ -18,20 +21,19 @@ public class DeleteNotePermanentlyCallback extends MaterialDialog.ButtonCallback
     }
 
     @Override
-    public void onNegative(MaterialDialog dialog) {
-        mAdapter.addItem(mPosition, mNote);
-        mAdapter.getRecyclerView().smoothScrollToPosition(mPosition);
-        dialog.cancel();
-    }
-
-    @Override
-    public void onPositive(MaterialDialog dialog) {
-        if (!MyApplication.getWritableDatabase().deleteNoteFromBin(mNote)) {
-            //failed to delete note from db
+    public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
+        if(which == DialogAction.POSITIVE) {
+            //positive option selected, delete note
+            if (!MyApplication.getWritableDatabase().deleteNoteFromBin(mNote)) {
+                //failed to delete note from db
+                mAdapter.addItem(mPosition, mNote);
+                mAdapter.getRecyclerView().smoothScrollToPosition(mPosition);
+                MyDebugger.log("Failed to delete note permanently!");
+            }
+        } else if(which == DialogAction.NEGATIVE) {
+            //negative option selected, bring note back
             mAdapter.addItem(mPosition, mNote);
             mAdapter.getRecyclerView().smoothScrollToPosition(mPosition);
-            MyDebugger.log("Failed to delete note permanently!");
         }
-        dialog.cancel();
     }
 }
